requirement.txt
Django==1.9.5
Pillow==3.2.0
argparse==1.2.1
django-admin-sortable2==0.6.3
django-banner>=0.2
django-ckeditor==5.0.3
django-flat-theme==1.1.3
django-mptt==0.8.3
wsgiref==0.1.2
